+++ jax_galsim/__init__.py
@@ -1,7 +1,47 @@
-from jax_galsim.exponential import Exponential
-from jax_galsim.gaussian import Gaussian
-from jax_galsim.sum import Add, Sum
+# First some basic building blocks that don't usually depend on anything else
+from jax_galsim.position import Position, PositionD, PositionI
+from jax_galsim.bounds import Bounds, BoundsI, BoundsD
+from galsim.angle import radians, hours, degrees, arcmin, arcsec
+
+# Inherit all Exception and Warning classes from galsim
+from galsim.errors import GalSimError, GalSimRangeError, GalSimValueError
+from galsim.errors import GalSimKeyError, GalSimIndexError, GalSimNotImplementedError
+from galsim.errors import (
+    GalSimBoundsError,
+    GalSimUndefinedBoundsError,
+    GalSimImmutableError,
+)
+from galsim.errors import GalSimIncompatibleValuesError, GalSimSEDError, GalSimHSMError
+from galsim.errors import GalSimFFTSizeError
+from galsim.errors import GalSimConfigError, GalSimConfigValueError
+from galsim.errors import GalSimWarning, GalSimDeprecationWarning
+
+# Image
+from .image import (
+    Image,
+    ImageS,
+    ImageI,
+    ImageF,
+    ImageD,
+    ImageCF,
+    ImageCD,
+    ImageUS,
+    ImageUI,
+)
+
+# GSObject
 from jax_galsim.gsobject import GSObject
 from jax_galsim.gsparams import GSParams
 
+
+from jax_galsim.gaussian import Gaussian
+from jax_galsim.exponential import Exponential
+from jax_galsim.sum import Add, Sum
+from jax_galsim.transform import Transformation, Transform
+
+# WCS
+from jax_galsim.wcs import PixelScale, OffsetWCS, JacobianWCS, AffineTransform
+
+# Noise
 from jax_galsim.noise import GaussianNoise
+