+++ CHANGELOG.md
@@ -4,3 +4,13 @@
 
 
 * Changes
+  * Added simplified versions of basic objects, with no support of Chromatic objects yet:
+    * `GSObjects` and `gsparams`
+    * `Sum`
+    * `Position`
+    * `Transformation`
+  * Added implementation of simple light profiles:
+    * `Gaussian`, `Exponential`
+  * Added implementation of simple WCS:
+    * `PixelScale`, `OffsetWCS`, `JacobianWCS`, `AffineTransform`
+  * Added automated suite of tests against reference GalSim